% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{For_xyzi}
\alias{For_xyzi}
\title{Extract XYZ Coordinates and Intensity from Point Cloud Data}
\usage{
For_xyzi(data)
}
\arguments{
\item{data}{A data frame containing point cloud data with columns named "X", "Y", "Z", and "Intensity".}
}
\value{
A matrix with four columns: X, Y, Z, and Intensity.
}
\description{
This function extracts the X, Y, Z coordinates and intensity values from a point cloud data frame and returns them as a matrix.
}
\examples{
pc_data <- data.frame(X = runif(10), Y = runif(10), Z = runif(10), Intensity = runif(10))
For_xyzi(pc_data)
}
