% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/architecture.R
\name{vectorized_custom_fun_bis}
\alias{vectorized_custom_fun_bis}
\title{Custom Vectorized Overlay Function}
\usage{
vectorized_custom_fun_bis(x, y, valx, valy)
}
\arguments{
\item{x}{A numeric vector representing values from the first raster layer.}

\item{y}{A numeric vector representing values from the second raster layer.}

\item{valx}{A numeric value to compare against elements in \code{x}.}

\item{valy}{A numeric value to compare against elements in \code{y}.}
}
\value{
A numeric vector where elements are \code{x + 1} if \code{x == valx} and \code{y == valy}, otherwise 0.
}
\description{
This function performs a conditional check on two raster layer values and returns a modified value based on the comparison with given constants.
}
\details{
This function is designed to be used within raster overlay operations. It is vectorized and can be applied efficiently to raster data using functions like \code{\link[raster]{overlay}}.
}
\examples{
x <- c(1, 2, 3, 4)
y <- c(5, 2, 3, 4)
vectorized_custom_fun_bis(x, y, valx = 3, valy = 3)
# Returns: 0 0 4 0
}
